[F-001] Mybatis 적용
[F-000] Mybtis 적용
- AppConfig.class @MapperScan 누락으로 인한 오류 수정
