+++ bin/main/dtd/mybatis-3-config.dtd
@@ -1,0 +1,117 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!--
+
+       Copyright 2009-2023 the original author or authors.
+
+       Licensed under the Apache License, Version 2.0 (the "License");
+       you may not use this file except in compliance with the License.
+       You may obtain a copy of the License at
+
+          https://www.apache.org/licenses/LICENSE-2.0
+
+       Unless required by applicable law or agreed to in writing, software
+       distributed under the License is distributed on an "AS IS" BASIS,
+       WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+       See the License for the specific language governing permissions and
+       limitations under the License.
+
+-->
+<!ELEMENT configuration (properties?, settings?, typeAliases?, typeHandlers?, objectFactory?, objectWrapperFactory?, reflectorFactory?, plugins?, environments?, databaseIdProvider?, mappers?)>
+
+<!ELEMENT databaseIdProvider (property*)>
+<!ATTLIST databaseIdProvider
+type CDATA #REQUIRED
+>
+
+<!ELEMENT properties (property*)>
+<!ATTLIST properties
+resource CDATA #IMPLIED
+url CDATA #IMPLIED
+>
+
+<!ELEMENT property EMPTY>
+<!ATTLIST property
+name CDATA #REQUIRED
+value CDATA #REQUIRED
+>
+
+<!ELEMENT settings (setting+)>
+
+<!ELEMENT setting EMPTY>
+<!ATTLIST setting
+name CDATA #REQUIRED
+value CDATA #REQUIRED
+>
+
+<!ELEMENT typeAliases (typeAlias*,package*)>
+
+<!ELEMENT typeAlias EMPTY>
+<!ATTLIST typeAlias
+type CDATA #REQUIRED
+alias CDATA #IMPLIED
+>
+
+<!ELEMENT typeHandlers (typeHandler*,package*)>
+
+<!ELEMENT typeHandler EMPTY>
+<!ATTLIST typeHandler
+javaType CDATA #IMPLIED
+jdbcType CDATA #IMPLIED
+handler CDATA #REQUIRED
+>
+
+<!ELEMENT objectFactory (property*)>
+<!ATTLIST objectFactory
+type CDATA #REQUIRED
+>
+
+<!ELEMENT objectWrapperFactory EMPTY>
+<!ATTLIST objectWrapperFactory
+type CDATA #REQUIRED
+>
+
+<!ELEMENT reflectorFactory EMPTY>
+<!ATTLIST reflectorFactory
+type CDATA #REQUIRED
+>
+
+<!ELEMENT plugins (plugin+)>
+
+<!ELEMENT plugin (property*)>
+<!ATTLIST plugin
+interceptor CDATA #REQUIRED
+>
+
+<!ELEMENT environments (environment+)>
+<!ATTLIST environments
+default CDATA #REQUIRED
+>
+
+<!ELEMENT environment (transactionManager,dataSource)>
+<!ATTLIST environment
+id CDATA #REQUIRED
+>
+
+<!ELEMENT transactionManager (property*)>
+<!ATTLIST transactionManager
+type CDATA #REQUIRED
+>
+
+<!ELEMENT dataSource (property*)>
+<!ATTLIST dataSource
+type CDATA #REQUIRED
+>
+
+<!ELEMENT mappers (mapper*,package*)>
+
+<!ELEMENT mapper EMPTY>
+<!ATTLIST mapper
+resource CDATA #IMPLIED
+url CDATA #IMPLIED
+class CDATA #IMPLIED
+>
+
+<!ELEMENT package EMPTY>
+<!ATTLIST package
+name CDATA #REQUIRED
+>
