+++ bin/main/mapper/UserMapper.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.example.mapper.UserMapper">
+	
+	<!-- 사용자 정보 조회 -->
+	<select id="getUserByEmail" parameterType="String" resultType="com.example.model.User">
+		SELECT ID
+			 , NAME
+			 , EMAIL
+			 , PASSWORD 
+		  FROM USERS
+		 WHERE email = #{email}
+	</select>
+</mapper>
