+++ build.gradle
@@ -12,7 +12,10 @@ repositories {
 }
 
 dependencies {
-    implementation 'org.apache.commons:commons-dbcp2:2.9.0'
+	//MyBatis 적용 (25.03.09)
+	implementation 'org.mybatis:mybatis:3.5.13'
+    implementation 'org.mybatis:mybatis-spring:2.0.7'
+	
     implementation 'org.springframework:spring-webmvc:5.3.30'
     implementation 'org.springframework:spring-jdbc:5.3.30'
     providedCompile 'javax.servlet:javax.servlet-api:4.0.1'
