+++ src/main/java/com/example/config/AppConfig.java
@@ -1,0 +1,72 @@
+package com.example.config;
+
+import javax.sql.DataSource;
+
+import org.apache.ibatis.datasource.pooled.PooledDataSource;
+import org.apache.ibatis.session.SqlSessionFactory;
+import org.mybatis.spring.SqlSessionFactoryBean;
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
+import org.springframework.web.servlet.ViewResolver;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+
+/**
+ * Application의 전반적인 설정을 관리하는 클래스.
+ * ## 01. Database 설정
+ * ## 02. MyBatis 설정
+ * ## 03. ViewResolver 설정 (JSP 경로 지정)
+ */
+@Configuration
+@EnableWebMvc
+@ComponentScan(basePackages = "com.example")
+@MapperScan(basePackages = "com.example.mapper") // MyBatis 매퍼 인터페이스 스캔
+public class AppConfig implements WebMvcConfigurer {
+
+	
+    /**
+     * Database 설정 (DataSource)
+     */
+    @Bean
+    public DataSource dataSource() {
+        PooledDataSource dataSource = new PooledDataSource();
+        dataSource.setDriver("org.postgresql.Driver");
+        dataSource.setUrl("jdbc:postgresql://localhost:5432/mydatabase");
+        dataSource.setUsername("postgres");
+        dataSource.setPassword("1234");
+        return dataSource;
+    }
+
+    /**
+     * MyBatis 설정 (SqlSessionFactory)
+     */
+    @Bean
+    public SqlSessionFactory sqlSessionFactory() throws Exception {
+        SqlSessionFactoryBean sessionFactory = new SqlSessionFactoryBean();
+
+        sessionFactory.setDataSource(dataSource());
+        
+        sessionFactory.setMapperLocations(
+        		new PathMatchingResourcePatternResolver().getResources("classpath:mapper/*.xml")
+        );
+        
+        return sessionFactory.getObject();
+    }
+
+    /**
+     * ViewResolver 설정 (JSP 파일 경로)
+     */
+    @Bean
+    public ViewResolver viewResolver() {
+        InternalResourceViewResolver resolver = new InternalResourceViewResolver();
+        resolver.setPrefix("/WEB-INF/views/");
+        resolver.setSuffix(".jsp");
+        return resolver;
+    }
+}
+
+
