--- src/main/java/com/example/config/DatabaseConfig.java
@@ -1,24 +1,0 @@
-package com.example.config;
-
-import javax.sql.DataSource;
-
-import org.apache.commons.dbcp2.BasicDataSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-
-@Configuration
-public class DatabaseConfig {
-	
-	@Bean
-	public DataSource dataSource() {
-		BasicDataSource dataSource = new BasicDataSource();
-		dataSource.setDriverClassName("org.postgresql.Driver");
-        dataSource.setUrl("jdbc:postgresql://localhost:5432/mydatabase");
-        dataSource.setUsername("postgres");
-        dataSource.setPassword("1234");
-        dataSource.setInitialSize(5);
-        dataSource.setMaxTotal(10);
-        return dataSource;
-	}
-}
