+++ src/main/java/com/example/config/WebAppInitializer.java
@@ -2,16 +2,21 @@ package com.example.config;
 
 import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
 
+/**
+ * Spring MVC 초기 설정 담당 클래스.
+ * web.xml 대체하는 역할 수행
+ * Spring 실행 시, DispatcherSevlet 등록, 설정 파일 로드 기능 수행
+ */
 public class WebAppInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {
 
     @Override
     protected Class<?>[] getRootConfigClasses() {
-        return new Class[]{DatabaseConfig.class};
+        return new Class[]{AppConfig.class}; // 모든 설정을 AppConfig에서 로드
     }
 
     @Override
     protected Class<?>[] getServletConfigClasses() {
-        return new Class[]{WebMvcConfig.class};
+        return null;
     }
 
     @Override
